Clean global context by creating hook
